requirements.txt
cd  fastapi
uvicorn
python-dotenv
python-multipart

langchain
langchain-core
langchain-community
langchain-google-genai
langchain-text-splitters

faiss-cpu
pypdf

tiktoken
python-multipart
sentence-transformers


python-docx
pdfplumber
langchain-community
langchain-huggingface
google-generativeai
